chore(ssi): move integration test file
This commit moves the high level tests into place in our module. I
originally put these in their own file because I wanted the git diff to
be understandable.
