+++ pkg/clusteragent/admission/mutate/autoinstrumentation/auto_instrumentation_test.go
@@ -6,3 +6,3095 @@
 //go:build kubeapiserver
 
 package autoinstrumentation_test
+
+import (
+	"fmt"
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/ptr"
+
+	"github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/util"
+	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
+	workloadmetamock "github.com/DataDog/datadog-agent/comp/core/workloadmeta/mock"
+	admissioncommon "github.com/DataDog/datadog-agent/pkg/clusteragent/admission/common"
+	"github.com/DataDog/datadog-agent/pkg/clusteragent/admission/mutate/autoinstrumentation"
+	"github.com/DataDog/datadog-agent/pkg/clusteragent/admission/mutate/common"
+	"github.com/DataDog/datadog-agent/pkg/languagedetection/languagemodels"
+)
+
+const (
+	defaultInjectorVersion = "0"
+	defaultTestContainer   = "test-container"
+)
+
+var defaultContainerNames = []string{defaultTestContainer}
+
+var defaultLibraries = map[string]string{
+	"dotnet": "v3",
+	"java":   "v1",
+	"js":     "v5",
+	"php":    "v1",
+	"python": "v3",
+	"ruby":   "v2",
+}
+
+// TestAutoinstrumentation is an integration style test that ensures user configuration maps to the expected pod
+// mutation.
+func TestAutoinstrumentation(t *testing.T) {
+	// expected is a struct that must be defined if should mutate is true. If injection is expected, one of the two
+	// options should be set (libraryVersions + injectorVersion) OR (initContainerImages).
+	type expected struct {
+		// injectorVersion (required) ensures the version of the injector matches the expected version.
+		injectorVersion string
+		// libraryVersions (required) ensures the versions of libraries expected exist in the pod.
+		libraryVersions map[string]string
+		// containerNames (required) is the list of containers where injection is expected.
+		containerNames []string
+		// initContainerImages (optional) ensures that the list of init container images exist in the pod rather then
+		// relying on injectorVersion and libraryVersions. This is useful when you also need to test the registry or
+		// custom image.
+		initContainerImages []string
+		// requiredEnvs (optional) ensures that the additional environment variables exist on every container.
+		requiredEnvs map[string]string
+		// unsetEnvs (optional) ensures that the environment variables do not exist on any container.
+		unsetEnvs []string
+		// initSecurityContext (optional) ensures that the init containers contain the additional security context.
+		initSecurityContext *corev1.SecurityContext
+		// initResourceRequirements (optional) ensures that the init containers contain the proper resource requirements.
+		initResourceRequirements *corev1.ResourceRequirements
+	}
+
+	tests := map[string]struct {
+		config       map[string]any
+		pod          *corev1.Pod
+		namespaces   []workloadmeta.KubernetesMetadata
+		deployments  []common.MockDeployment
+		shouldMutate bool
+		expected     *expected
+	}{
+		"default configuration should not mutate": {
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod with annotation but without mutate label and mutate unlabelled disabled should not mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"local lib injection should set install type": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+				requiredEnvs: map[string]string{
+					"DD_INSTRUMENTATION_INSTALL_TYPE": "k8s_lib_injection",
+				},
+			},
+		},
+		"pod with mutate label and java annotation should mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with mutate label and python annotation should mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/python-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"python": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with mutate label and js annotation should mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/js-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"js": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with mutate label and php annotation should mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/php-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"php": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with mutate label and ruby annotation should mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/ruby-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"ruby": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with mutate label and dotnet annotation should mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/dotnet-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"dotnet": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with local sdk injection debug enabled should get debug info": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+						"admission.datadoghq.com/apm-inject.debug": "true",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				requiredEnvs: map[string]string{
+					"DD_APM_INSTRUMENTATION_DEBUG": "true",
+					"DD_TRACE_STARTUP_LOGS":        "true",
+					"DD_TRACE_DEBUG":               "true",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with local sdk injection debug disabled should not get debug info": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+						"admission.datadoghq.com/apm-inject.debug": "false",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				unsetEnvs: []string{
+					"DD_APM_INSTRUMENTATION_DEBUG",
+					"DD_TRACE_STARTUP_LOGS",
+					"DD_TRACE_DEBUG",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with local sdk injection debug enabled but no library should not get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/apm-inject.debug": "true",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod with mutate label disabled and annotation should not mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "false",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod with mutate label disabled with mutate unlabelled and annotation should not mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "false",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod with mutate label but no annotation and instrumentation disabled should not mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod with mutate label but no annotation and instrumentation enabled should get default libs": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     true,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"pod without mutate label and mutate unlabelled enabled should mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with annotation should take precedence over config": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.lib_versions": map[string]string{
+					"java": "v2",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with instrumentation enabled should get all default libs": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"pod with instrumentation enabled and lib versions defined should get defined versions": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.lib_versions": map[string]string{
+					"python": "v5",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"python": "v5",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with instrumentation enabled and lib versions defined with a magic tag should get resolved": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.lib_versions": map[string]string{
+					"java": "default",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod outside of enabled namespace should not get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.enabled_namespaces": []string{
+					"foo",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod inside of enabled namespace should get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.enabled_namespaces": []string{
+					"application",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"pod outside of enabled namespace but with local lib injection should get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.enabled_namespaces": []string{
+					"foo",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod inside of disabled namespace should not get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.disabled_namespaces": []string{
+					"application",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod outside of disabled namespace should get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.disabled_namespaces": []string{
+					"foo",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"pod inside of disabled namespace but with local lib injection should not get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.disabled_namespaces": []string{
+					"application",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod inside of disabled namespace but with local lib injection and instrumentation disabled should not get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": false,
+				"apm_config.instrumentation.disabled_namespaces": []string{
+					"application",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod with local lib injection and default tag should resolve to version": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "default",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with local lib injection that defines two libs should get both": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+						"admission.datadoghq.com/js-lib.version":   "v3",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+					"js":   "v3",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with local lib injection that sets injector version should get it": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version":   "v1",
+						"admission.datadoghq.com/apm-inject.version": "v3",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: "v3",
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with local lib injection that defines an unknown version should not mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/unknown-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod with local lib injection using all annotation gets all libs": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/all-lib.version": "latest",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"pod with local lib injection using all annotation but no label does not get mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/all-lib.version": "latest",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"pod with local lib injection using all annotation, no label, and mutated unlabelled gets mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/all-lib.version": "latest",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"pod with local lib injection using pinned versions gets mutated": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1.2.3",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1.2.3",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with local lib injection using java and all annotations only gets the java image": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version": "v1",
+						"admission.datadoghq.com/all-lib.version":  "latest",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with local lib injection using all annotations with unsupported tag still gets all versions": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/all-lib.version": "unsupported",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"pod with local lib injection and custom injector gets custom version": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":            false,
+				"admission_controller.mutate_unlabelled":        false,
+				"apm_config.instrumentation.injector_image_tag": "1.2.3",
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/php-lib.version": "5.2.1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: "1.2.3",
+				libraryVersions: map[string]string{
+					"php": "5.2.1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"pod with enabled namespaces and custom injector gets custom version": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":            true,
+				"apm_config.instrumentation.injector_image_tag": "1.2.3",
+				"apm_config.instrumentation.enabled_namespaces": []string{
+					"application",
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: "1.2.3",
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"targets with matching rule mutates pod": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.targets": []autoinstrumentation.Target{
+					{
+						Name: "test-target",
+						TracerVersions: map[string]string{
+							"ruby": "v2",
+						},
+						NamespaceSelector: &autoinstrumentation.NamespaceSelector{
+							MatchNames: []string{
+								"application",
+							},
+						},
+					},
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"ruby": "v2",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"targets with matching rule sets install type": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.targets": []autoinstrumentation.Target{
+					{
+						Name: "test-target",
+						TracerVersions: map[string]string{
+							"ruby": "v2",
+						},
+						NamespaceSelector: &autoinstrumentation.NamespaceSelector{
+							MatchNames: []string{
+								"application",
+							},
+						},
+					},
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"ruby": "v2",
+				},
+				containerNames: defaultContainerNames,
+				requiredEnvs: map[string]string{
+					"DD_INSTRUMENTATION_INSTALL_TYPE": "k8s_single_step",
+				},
+			},
+		},
+		"targets with matching rule and additional config both apply to pod": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.targets": []autoinstrumentation.Target{
+					{
+						Name: "test-target",
+						TracerVersions: map[string]string{
+							"ruby": "v2",
+						},
+						TracerConfigs: []autoinstrumentation.TracerConfig{
+							{
+								Name:  "DD_PROFILING_ENABLED",
+								Value: "true",
+							},
+						},
+						NamespaceSelector: &autoinstrumentation.NamespaceSelector{
+							MatchNames: []string{
+								"application",
+							},
+						},
+					},
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"ruby": "v2",
+				},
+				requiredEnvs: map[string]string{
+					"DD_PROFILING_ENABLED": "true",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"targets without matching rule does not mutate pod": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.targets": []autoinstrumentation.Target{
+					{
+						Name: "test-target",
+						NamespaceSelector: &autoinstrumentation.NamespaceSelector{
+							MatchNames: []string{
+								"foo",
+							},
+						},
+					},
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"targets with matching rule and local sdk injection favors local sdk version": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.targets": []autoinstrumentation.Target{
+					{
+						Name: "test-target",
+						TracerVersions: map[string]string{
+							"ruby": "v2",
+						},
+						NamespaceSelector: &autoinstrumentation.NamespaceSelector{
+							MatchNames: []string{
+								"application",
+							},
+						},
+					},
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/ruby-lib.version": "v3",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"ruby": "v3",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"targets with matching rule and local sdk injection but no label favors target": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.targets": []autoinstrumentation.Target{
+					{
+						Name: "test-target",
+						TracerVersions: map[string]string{
+							"ruby": "v2",
+						},
+						NamespaceSelector: &autoinstrumentation.NamespaceSelector{
+							MatchNames: []string{
+								"application",
+							},
+						},
+					},
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/ruby-lib.version": "v3",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"ruby": "v2",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"targets with matching rule and local sdk injection with no label but with mutate unlabelled favors local sdk": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     true,
+				"admission_controller.mutate_unlabelled": true,
+				"apm_config.instrumentation.targets": []autoinstrumentation.Target{
+					{
+						Name: "test-target",
+						TracerVersions: map[string]string{
+							"ruby": "v2",
+						},
+						NamespaceSelector: &autoinstrumentation.NamespaceSelector{
+							MatchNames: []string{
+								"application",
+							},
+						},
+					},
+				},
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/ruby-lib.version": "v3",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"ruby": "v3",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"local sdk injection with custom injector image gets custom image": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/php-lib.version":         "v1",
+						"admission.datadoghq.com/apm-inject.custom-image": "docker.io/library/apm-inject-package:v27",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				initContainerImages: []string{
+					"gcr.io/datadoghq/dd-lib-php-init:v1",
+					"docker.io/library/apm-inject-package:v27",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"local sdk injection with custom library image gets custom image": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/php-lib.custom-image": "foo/bar:1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				initContainerImages: []string{
+					"gcr.io/datadoghq/apm-inject:0",
+					"foo/bar:1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"local sdk injection with configs get configs set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.version":   "v1",
+						"admission.datadoghq.com/java-lib.config.v1": `{"runtime_metrics_enabled":true,"tracing_rate_limit":50}`,
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				requiredEnvs: map[string]string{
+					"DD_RUNTIME_METRICS_ENABLED": "true",
+					"DD_TRACE_RATE_LIMIT":        "50",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"instrumentation enabled and no mutate label with config annotation gets configs set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"apm_config.instrumentation.lib_versions": map[string]string{
+					"java": "v2",
+				},
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/java-lib.config.v1": `{"runtime_metrics_enabled":true,"tracing_rate_limit":50}`,
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v2",
+				},
+				requiredEnvs: map[string]string{
+					"DD_RUNTIME_METRICS_ENABLED": "true",
+					"DD_TRACE_RATE_LIMIT":        "50",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"local sdk injection with all lib get configs set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/all-lib.version":   "latest",
+						"admission.datadoghq.com/all-lib.config.v1": `{"runtime_metrics_enabled":true,"tracing_rate_limit":50,"tracing_sampling_rate":0.3}`,
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: "test-container",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				requiredEnvs: map[string]string{
+					"DD_RUNTIME_METRICS_ENABLED": "true",
+					"DD_TRACE_RATE_LIMIT":        "50",
+					"DD_TRACE_SAMPLE_RATE":       "0.30",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"istio-proxy is not injected": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+						},
+						{
+							Name: "istio-proxy",
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames: []string{
+					defaultTestContainer,
+				},
+			},
+		},
+		"injection does not occur in the namespace where datadog is deployed": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"kube_resources_namespace":           "datadog-test",
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "datadog-test",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+						},
+					},
+				},
+			},
+			shouldMutate: false,
+		},
+		"injection does occur in the outside the datadog namespace": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+				"kube_resources_namespace":           "datadog-test",
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+			},
+		},
+		"language detection enabled limits the injected libraries": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":                                       true,
+				"language_detection.enabled":                                               true,
+				"language_detection.reporting.enabled":                                     true,
+				"admission_controller.auto_instrumentation.inject_auto_detected_libraries": true,
+			},
+			pod: common.FakePodSpec{
+				Name:       defaultTestContainer,
+				NS:         "application",
+				ParentKind: "replicaset",
+				ParentName: "deployment-123",
+			}.Create(),
+			deployments: []common.MockDeployment{
+				{
+					ContainerName:  defaultTestContainer,
+					DeploymentName: "deployment",
+					Namespace:      "application",
+					Languages:      languageSetOf("python"),
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"python": defaultLibraries["python"],
+				},
+				containerNames: defaultContainerNames,
+				requiredEnvs: map[string]string{
+					"DD_INSTRUMENTATION_LANGUAGES_DETECTED":                   "python",
+					"DD_INSTRUMENTATION_LANGUAGE_DETECTION_INJECTION_ENABLED": "true",
+				},
+			},
+		},
+		"language detection disabled sets disabled env var": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":                                       true,
+				"language_detection.enabled":                                               true,
+				"language_detection.reporting.enabled":                                     true,
+				"admission_controller.auto_instrumentation.inject_auto_detected_libraries": false,
+			},
+			pod: common.FakePodSpec{
+				Name:       defaultTestContainer,
+				NS:         "application",
+				ParentKind: "replicaset",
+				ParentName: "deployment-123",
+			}.Create(),
+			deployments: []common.MockDeployment{
+				{
+					ContainerName:  defaultTestContainer,
+					DeploymentName: "deployment",
+					Namespace:      "application",
+					Languages:      languageSetOf("python"),
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+				requiredEnvs: map[string]string{
+					"DD_INSTRUMENTATION_LANGUAGE_DETECTION_INJECTION_ENABLED": "false",
+				},
+			},
+		},
+		"language detection enabled but no language found does not limit libraries": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":                                       true,
+				"language_detection.enabled":                                               true,
+				"language_detection.reporting.enabled":                                     true,
+				"admission_controller.auto_instrumentation.inject_auto_detected_libraries": true,
+			},
+			pod: common.FakePodSpec{
+				Name:       defaultTestContainer,
+				NS:         "application",
+				ParentKind: "replicaset",
+				ParentName: "deployment-123",
+			}.Create(),
+			deployments: []common.MockDeployment{
+				{
+					ContainerName:  defaultTestContainer,
+					DeploymentName: "deployment",
+					Namespace:      "application",
+					Languages:      languageSetOf(),
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+				requiredEnvs: map[string]string{
+					"DD_INSTRUMENTATION_LANGUAGES_DETECTED":                   "",
+					"DD_INSTRUMENTATION_LANGUAGE_DETECTION_INJECTION_ENABLED": "true",
+				},
+			},
+		},
+		"local lib injection per container should mutate": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":     false,
+				"admission_controller.mutate_unlabelled": false,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+					Annotations: map[string]string{
+						"admission.datadoghq.com/test-container.java-lib.version": "v1",
+					},
+					Labels: map[string]string{
+						admissioncommon.EnabledLabelKey: "true",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: map[string]string{
+					"java": "v1",
+				},
+				containerNames: defaultContainerNames,
+			},
+		},
+		"when a security context is set through config, it is applied": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":                              true,
+				"admission_controller.auto_instrumentation.init_security_context": `{"privileged":true}`,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+				initSecurityContext: &corev1.SecurityContext{
+					Privileged: ptr.To(true),
+				},
+			},
+		},
+		"when a pod is created in a privileged namespace, the default security context is set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+						},
+					},
+				},
+			},
+			namespaces: []workloadmeta.KubernetesMetadata{
+				newTestNamespace("application", map[string]string{
+					"pod-security.kubernetes.io/enforce": "restricted",
+				}),
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+				initSecurityContext: &corev1.SecurityContext{
+					AllowPrivilegeEscalation: ptr.To(false),
+					RunAsNonRoot:             ptr.To(true),
+					SeccompProfile:           &corev1.SeccompProfile{Type: corev1.SeccompProfileTypeRuntimeDefault},
+					Capabilities: &corev1.Capabilities{
+						Drop: []corev1.Capability{
+							"ALL",
+						},
+					},
+				},
+			},
+		},
+		"a pod with resources has init container resources set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("100m"),
+									corev1.ResourceMemory: resource.MustParse("500Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("100m"),
+									corev1.ResourceMemory: resource.MustParse("500Mi"),
+								},
+							},
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+				initResourceRequirements: &corev1.ResourceRequirements{
+					Limits: corev1.ResourceList{
+						corev1.ResourceCPU:    resource.MustParse("100m"),
+						corev1.ResourceMemory: resource.MustParse("500Mi"),
+					},
+					Requests: corev1.ResourceList{
+						corev1.ResourceCPU:    resource.MustParse("100m"),
+						corev1.ResourceMemory: resource.MustParse("500Mi"),
+					},
+				},
+			},
+		},
+		"a pod with only cpu resources has only cpu init container resources set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU: resource.MustParse("100m"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU: resource.MustParse("100m"),
+								},
+							},
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+				initResourceRequirements: &corev1.ResourceRequirements{
+					Limits: corev1.ResourceList{
+						corev1.ResourceCPU: resource.MustParse("100m"),
+					},
+					Requests: corev1.ResourceList{
+						corev1.ResourceCPU: resource.MustParse("100m"),
+					},
+				},
+			},
+		},
+		"a pod with only memory resources has only memory init container resources set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceMemory: resource.MustParse("500Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceMemory: resource.MustParse("500Mi"),
+								},
+							},
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+				initResourceRequirements: &corev1.ResourceRequirements{
+					Limits: corev1.ResourceList{
+						corev1.ResourceMemory: resource.MustParse("500Mi"),
+					},
+					Requests: corev1.ResourceList{
+						corev1.ResourceMemory: resource.MustParse("500Mi"),
+					},
+				},
+			},
+		},
+		"a pod with init container resources has the highest value init container resources set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					InitContainers: []corev1.Container{
+						{
+							Name: "foo",
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("700m"),
+									corev1.ResourceMemory: resource.MustParse("500Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("700m"),
+									corev1.ResourceMemory: resource.MustParse("500Mi"),
+								},
+							},
+						},
+						{
+							Name: "bar",
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("101m"),
+									corev1.ResourceMemory: resource.MustParse("700Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("101m"),
+									corev1.ResourceMemory: resource.MustParse("700Mi"),
+								},
+							},
+						},
+					},
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("102m"),
+									corev1.ResourceMemory: resource.MustParse("500Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("102m"),
+									corev1.ResourceMemory: resource.MustParse("500Mi"),
+								},
+							},
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames:  defaultContainerNames,
+				initResourceRequirements: &corev1.ResourceRequirements{
+					Limits: corev1.ResourceList{
+						corev1.ResourceCPU:    resource.MustParse("700m"),
+						corev1.ResourceMemory: resource.MustParse("700Mi"),
+					},
+					Requests: corev1.ResourceList{
+						corev1.ResourceCPU:    resource.MustParse("700m"),
+						corev1.ResourceMemory: resource.MustParse("700Mi"),
+					},
+				},
+			},
+		},
+		"a pod with multiple container resources is additive": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					InitContainers: []corev1.Container{
+						{
+							Name: "foo",
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("100m"),
+									corev1.ResourceMemory: resource.MustParse("100Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("100m"),
+									corev1.ResourceMemory: resource.MustParse("100Mi"),
+								},
+							},
+						},
+					},
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("200m"),
+									corev1.ResourceMemory: resource.MustParse("200Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("200m"),
+									corev1.ResourceMemory: resource.MustParse("200Mi"),
+								},
+							},
+						},
+						{
+							Name: "sidecar",
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("200m"),
+									corev1.ResourceMemory: resource.MustParse("200Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("200m"),
+									corev1.ResourceMemory: resource.MustParse("200Mi"),
+								},
+							},
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames: []string{
+					defaultTestContainer,
+					"sidecar",
+				},
+				initResourceRequirements: &corev1.ResourceRequirements{
+					Limits: corev1.ResourceList{
+						corev1.ResourceCPU:    resource.MustParse("400m"),
+						corev1.ResourceMemory: resource.MustParse("400Mi"),
+					},
+					Requests: corev1.ResourceList{
+						corev1.ResourceCPU:    resource.MustParse("400m"),
+						corev1.ResourceMemory: resource.MustParse("400Mi"),
+					},
+				},
+			},
+		},
+		"when config is set, it takes precedence for resources": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":                              true,
+				"admission_controller.auto_instrumentation.init_resources.cpu":    "101m",
+				"admission_controller.auto_instrumentation.init_resources.memory": "301Mi",
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					InitContainers: []corev1.Container{
+						{
+							Name: "foo",
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("100m"),
+									corev1.ResourceMemory: resource.MustParse("100Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("100m"),
+									corev1.ResourceMemory: resource.MustParse("100Mi"),
+								},
+							},
+						},
+					},
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("200m"),
+									corev1.ResourceMemory: resource.MustParse("200Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("200m"),
+									corev1.ResourceMemory: resource.MustParse("200Mi"),
+								},
+							},
+						},
+						{
+							Name: "sidecar",
+							Resources: corev1.ResourceRequirements{
+								Limits: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("200m"),
+									corev1.ResourceMemory: resource.MustParse("200Mi"),
+								},
+								Requests: corev1.ResourceList{
+									corev1.ResourceCPU:    resource.MustParse("200m"),
+									corev1.ResourceMemory: resource.MustParse("200Mi"),
+								},
+							},
+						},
+					},
+				},
+			},
+			shouldMutate: true,
+			expected: &expected{
+				injectorVersion: defaultInjectorVersion,
+				libraryVersions: defaultLibraries,
+				containerNames: []string{
+					defaultTestContainer,
+					"sidecar",
+				},
+				initResourceRequirements: &corev1.ResourceRequirements{
+					Limits: corev1.ResourceList{
+						corev1.ResourceCPU:    resource.MustParse("101m"),
+						corev1.ResourceMemory: resource.MustParse("301Mi"),
+					},
+					Requests: corev1.ResourceList{
+						corev1.ResourceCPU:    resource.MustParse("101m"),
+						corev1.ResourceMemory: resource.MustParse("301Mi"),
+					},
+				},
+			},
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			// Setup mocks.
+			mockConfig := common.FakeConfigWithValues(t, test.config)
+			mockMeta := common.FakeStoreWithDeployment(t, test.deployments)
+			mockImageResolver := NewFakeImageResolver()
+
+			// Add the namespaces.
+			for _, ns := range test.namespaces {
+				mockMeta.(workloadmetamock.Mock).Set(&ns)
+			}
+
+			// Setup webhook.
+			config, err := autoinstrumentation.NewConfig(mockConfig)
+			require.NoError(t, err)
+			mutator, err := autoinstrumentation.NewTargetMutator(config, mockMeta, mockImageResolver)
+			require.NoError(t, err)
+			webhook, err := autoinstrumentation.NewWebhook(config.Webhook, mockMeta, mutator)
+			require.NoError(t, err)
+
+			// Mutate pod.
+			in := test.pod.DeepCopy()
+			mutated, err := webhook.MutatePod(in, in.Namespace, nil)
+			require.NoError(t, err)
+
+			// If no mutation is expected, the pod should be identical and the boolean returned must be false.
+			if !test.shouldMutate {
+				require.Nil(t, test.expected, "the test was not properly configured, no mutation should not have expectations")
+				require.Equal(t, test.pod, in, "the pod was mutated by the webhook when no mutation is expected")
+				require.False(t, mutated, "mutate pod should only return true if the pod is mutated")
+				return
+			}
+
+			// Require the test to be setup correctly and for mutation to have occurred.
+			require.NotNil(t, test.expected, "the test was not properly configured, mutation should have expectations")
+			require.NotEmpty(t, test.expected.containerNames, "the test was not properly configured, containerNames must be set")
+			require.NotEqual(t, test.pod, in, "the pod was not mutated when it was expected to be")
+			require.True(t, mutated, "the pod was mutated but the webhook returned false")
+
+			// Require injection to have occurred.
+			validator := newPodValidator(in)
+			validator.requireInjection(t, test.expected.containerNames)
+
+			// Require the libraries and versions to match.
+			if len(test.expected.initContainerImages) > 0 {
+				require.Empty(t, test.expected.libraryVersions, "the test was not properly configured, library versions should only be set without init container images")
+				require.Empty(t, test.expected.injectorVersion, "the test was not properly configured, injector version should only be set without init container images")
+				validator.requireInitContainerImages(t, test.expected.initContainerImages)
+			} else {
+				validator.requireInjectorVersion(t, test.expected.injectorVersion)
+				validator.requireLibraryVersions(t, test.expected.libraryVersions)
+			}
+
+			// Require environments to be set.
+			validator.requireEnvs(t, test.expected.requiredEnvs, test.expected.containerNames)
+			validator.requireMissingEnvs(t, test.expected.unsetEnvs, test.expected.containerNames)
+
+			// Require security context to match expected.
+			validator.requireInitSecurityContext(t, test.expected.initSecurityContext)
+
+			// Require resources match expected.
+			validator.requireInitResourceRequirements(t, test.expected.initResourceRequirements)
+		})
+	}
+}
+
+func TestEnvVarsAlreadySet(t *testing.T) {
+	tests := map[string]struct {
+		config             map[string]any
+		pod                *corev1.Pod
+		expected           map[string]string
+		expectedContainers []string
+	}{
+		"a pod without LD_PRELOAD has it set": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+						},
+					},
+				},
+			},
+			expected: map[string]string{
+				"LD_PRELOAD": "/opt/datadog-packages/datadog-apm-inject/stable/inject/launcher.preload.so",
+			},
+			expectedContainers: defaultContainerNames,
+		},
+		"a pod with LD_PRELOAD already set has it appended": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+							Env: []corev1.EnvVar{
+								{
+									Name:  "LD_PRELOAD",
+									Value: "/foo",
+								},
+							},
+						},
+					},
+				},
+			},
+			expected: map[string]string{
+				"LD_PRELOAD": "/foo:/opt/datadog-packages/datadog-apm-inject/stable/inject/launcher.preload.so",
+			},
+			expectedContainers: defaultContainerNames,
+		},
+		"a pod with several LD_PRELOAD already set has it appended": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+							Env: []corev1.EnvVar{
+								{
+									Name:  "LD_PRELOAD",
+									Value: "/foo:/bar",
+								},
+							},
+						},
+					},
+				},
+			},
+			expected: map[string]string{
+				"LD_PRELOAD": "/foo:/bar:/opt/datadog-packages/datadog-apm-inject/stable/inject/launcher.preload.so",
+			},
+			expectedContainers: defaultContainerNames,
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			// Setup mocks.
+			mockConfig := common.FakeConfigWithValues(t, test.config)
+			mockMeta := common.FakeStore(t)
+			mockImageResolver := NewFakeImageResolver()
+
+			// Setup webhook.
+			config, err := autoinstrumentation.NewConfig(mockConfig)
+			require.NoError(t, err)
+			mutator, err := autoinstrumentation.NewTargetMutator(config, mockMeta, mockImageResolver)
+			require.NoError(t, err)
+			webhook, err := autoinstrumentation.NewWebhook(config.Webhook, mockMeta, mutator)
+			require.NoError(t, err)
+
+			// Mutate pod.
+			in := test.pod.DeepCopy()
+			mutated, err := webhook.MutatePod(in, in.Namespace, nil)
+			require.NoError(t, err)
+			require.True(t, mutated, "the pod was mutated but the webhook returned false")
+
+			// Setup validator.
+			validator := newPodValidator(in)
+
+			// Require environment to match.
+			validator.requireEnvs(t, test.expected, test.expectedContainers)
+		})
+	}
+}
+
+func TestSkippedDueToResources(t *testing.T) {
+	tests := map[string]struct {
+		config              map[string]any
+		pod                 *corev1.Pod
+		skipped             bool
+		expectedContainers  []string
+		expectedAnnotations map[string]string
+	}{
+		"a pod with ample resources is not skipped": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: &corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "application",
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name: defaultTestContainer,
+						},
+					},
+				},
+			},
+			skipped:            false,
+			expectedContainers: defaultContainerNames,
+		},
+		"a pod with low memory is skipped": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: common.FakePodWithContainer("java-pod", corev1.Container{Name: defaultTestContainer, Resources: corev1.ResourceRequirements{
+				Limits: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("499m"),
+					corev1.ResourceMemory: resource.MustParse("50Mi"),
+				},
+				Requests: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("499m"),
+					corev1.ResourceMemory: resource.MustParse("50Mi"),
+				},
+			}}),
+			skipped:            true,
+			expectedContainers: defaultContainerNames,
+			expectedAnnotations: map[string]string{
+				"apm.datadoghq.com/injection-error": "The overall pod's containers limit is too low, memory pod_limit=50Mi needed=100Mi",
+			},
+		},
+		"a pod with low cpu is skipped": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: common.FakePodWithContainer("java-pod", corev1.Container{Name: defaultTestContainer, Resources: corev1.ResourceRequirements{
+				Limits: corev1.ResourceList{
+					corev1.ResourceCPU: resource.MustParse("0.025"),
+				},
+				Requests: corev1.ResourceList{
+					corev1.ResourceCPU: resource.MustParse("0.025"),
+				},
+			}}),
+			skipped:            true,
+			expectedContainers: defaultContainerNames,
+			expectedAnnotations: map[string]string{
+				"apm.datadoghq.com/injection-error": "The overall pod's containers limit is too low, cpu pod_limit=25m needed=50m",
+			},
+		},
+		"a pod with low cpu and memory is skipped": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled": true,
+			},
+			pod: common.FakePodWithContainer("java-pod", corev1.Container{Name: defaultTestContainer, Resources: corev1.ResourceRequirements{
+				Limits: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("0.025"),
+					corev1.ResourceMemory: resource.MustParse("50Mi"),
+				},
+				Requests: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("0.025"),
+					corev1.ResourceMemory: resource.MustParse("50Mi"),
+				},
+			}}),
+			skipped:            true,
+			expectedContainers: defaultContainerNames,
+			expectedAnnotations: map[string]string{
+				"apm.datadoghq.com/injection-error": "The overall pod's containers limit is too low, cpu pod_limit=25m needed=50m, memory pod_limit=50Mi needed=100Mi",
+			},
+		},
+		"a pod with low cpu and memory but with config override is not skipped": {
+			config: map[string]any{
+				"apm_config.instrumentation.enabled":                              true,
+				"admission_controller.auto_instrumentation.init_resources.cpu":    "101m",
+				"admission_controller.auto_instrumentation.init_resources.memory": "301Mi",
+			},
+			pod: common.FakePodWithContainer("java-pod", corev1.Container{Name: defaultTestContainer, Resources: corev1.ResourceRequirements{
+				Limits: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("0.025"),
+					corev1.ResourceMemory: resource.MustParse("50Mi"),
+				},
+				Requests: corev1.ResourceList{
+					corev1.ResourceCPU:    resource.MustParse("0.025"),
+					corev1.ResourceMemory: resource.MustParse("50Mi"),
+				},
+			}}),
+			skipped:            false,
+			expectedContainers: defaultContainerNames,
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			// Setup mocks.
+			mockConfig := common.FakeConfigWithValues(t, test.config)
+			mockMeta := common.FakeStore(t)
+			mockImageResolver := NewFakeImageResolver()
+
+			// Setup webhook.
+			config, err := autoinstrumentation.NewConfig(mockConfig)
+			require.NoError(t, err)
+			mutator, err := autoinstrumentation.NewTargetMutator(config, mockMeta, mockImageResolver)
+			require.NoError(t, err)
+			webhook, err := autoinstrumentation.NewWebhook(config.Webhook, mockMeta, mutator)
+			require.NoError(t, err)
+
+			// Mutate pod.
+			in := test.pod.DeepCopy()
+			mutated, err := webhook.MutatePod(in, in.Namespace, nil)
+			require.NoError(t, err)
+			require.True(t, mutated, "the pod was mutated but the webhook returned false")
+
+			// Setup validator.
+			validator := newPodValidator(in)
+
+			// Ensure the pod was properly skipped due to resources.
+			if test.skipped {
+				missingEnv := []string{
+					"LD_PRELOAD",
+				}
+				validator.requireMissingEnvs(t, missingEnv, test.expectedContainers)
+				validator.requireAnnotations(t, test.expectedAnnotations)
+				return
+			}
+
+			// Otherwise, require injection.
+			validator.requireInjection(t, test.expectedContainers)
+		})
+	}
+}
+
+type image struct {
+	raw      string
+	name     string
+	registry string
+	tag      string
+}
+
+func newImage(i string) *image {
+	// gcr.io/datadoghq/dd-lib-java-init:v1
+	parts := strings.Split(i, ":")
+	if len(parts) != 2 {
+		return nil
+	}
+
+	fullImage := parts[0]
+	tag := parts[1]
+
+	// gcr.io/datadoghq/dd-lib-java-init
+	parts = strings.Split(fullImage, "/")
+	if len(parts) < 1 {
+		return nil
+	}
+	name := parts[len(parts)-1]
+	registry := strings.TrimSuffix(fullImage, fmt.Sprintf("/%s", name))
+
+	return &image{
+		raw:      i,
+		name:     name,
+		registry: registry,
+		tag:      tag,
+	}
+}
+
+type podValidator struct {
+	raw                 *corev1.Pod
+	initContainers      map[string]*containerValidator
+	containers          map[string]*containerValidator
+	allContainerNames   []string
+	injectorVersion     string
+	libraryVersions     map[string]string
+	initContainerImages []string
+}
+
+func newPodValidator(pod *corev1.Pod) *podValidator {
+	containers := make(map[string]*containerValidator, len(pod.Spec.Containers))
+	for _, container := range pod.Spec.Containers {
+		containers[container.Name] = newContainerValidator(&container)
+	}
+
+	initContainers := make(map[string]*containerValidator, len(pod.Spec.InitContainers))
+	for _, container := range pod.Spec.InitContainers {
+		initContainers[container.Name] = newContainerValidator(&container)
+	}
+
+	return &podValidator{
+		raw:                 pod,
+		initContainers:      initContainers,
+		containers:          containers,
+		allContainerNames:   parseAllContainerNames(pod),
+		injectorVersion:     parseInjectorVersion(pod),
+		libraryVersions:     parseLibraryVersions(pod),
+		initContainerImages: parseInitContainerImages(pod),
+	}
+}
+
+func (v *podValidator) requireInitSecurityContext(t *testing.T, expected *corev1.SecurityContext) {
+	for _, validator := range v.initContainers {
+		validator.requireSecurityContext(t, expected)
+	}
+}
+
+func (v *podValidator) requireInitResourceRequirements(t *testing.T, expected *corev1.ResourceRequirements) {
+	for name, validator := range v.initContainers {
+		if !strings.HasPrefix(name, "datadog-") {
+			continue
+		}
+
+		validator.requireResourceRequirements(t, expected)
+	}
+}
+
+func (v *podValidator) requireInjection(t *testing.T, expectedContainers []string) {
+	// Validate the containers are injected that are expected to be.
+	for _, containerName := range expectedContainers {
+		validator, found := v.containers[containerName]
+		require.True(t, found, "invalid test setup, expected container %s does not exist in the pod", containerName)
+		validator.requireInjection(t)
+	}
+
+	// Validate the containers not expected to be injected are not.
+	notExpected := difference(v.allContainerNames, expectedContainers)
+	for _, containerName := range notExpected {
+		validator, found := v.containers[containerName]
+		require.True(t, found, "invalid test setup, expected container %s does not exist in the pod", containerName)
+		validator.requireNoInjection(t)
+	}
+
+	// Validate pod level volumes are created.
+	expectedVolumeNames := []string{
+		"datadog-auto-instrumentation",
+		"datadog-auto-instrumentation-etc",
+	}
+	v.requireVolumeNames(t, expectedVolumeNames)
+
+	// Validate pod annotations.
+	expectedAnnotations := map[string]string{
+		common.K8sAutoscalerSafeToEvictVolumesAnnotation: "datadog-auto-instrumentation,datadog-auto-instrumentation-etc",
+	}
+	v.requireAnnotations(t, expectedAnnotations)
+
+	// Validate injector init container.
+	validator, ok := v.initContainers["datadog-init-apm-inject"]
+	require.True(t, ok, "could not find datadog inject init container")
+	expectedVolumeMounts := []corev1.VolumeMount{
+		{
+			Name:      "datadog-auto-instrumentation",
+			MountPath: "/datadog-inject",
+			SubPath:   "opt/datadog-packages/datadog-apm-inject",
+		},
+		{
+			Name:      "datadog-auto-instrumentation-etc",
+			MountPath: "/datadog-etc",
+			SubPath:   "",
+		},
+	}
+	validator.requireVolumeMounts(t, expectedVolumeMounts)
+	validator.requireCommand(t, "/bin/sh -c -- cp -r /opt/datadog-packages/datadog-apm-inject/* /datadog-inject && echo /opt/datadog-packages/datadog-apm-inject/stable/inject/launcher.preload.so > /datadog-etc/ld.so.preload && echo $(date +%s) >> /datadog-inject/c-init-time.datadog-init-apm-inject")
+
+	// Validate library init containers.
+	for lang := range v.libraryVersions {
+		validator, ok := v.initContainers[fmt.Sprintf("datadog-lib-%s-init", lang)]
+		require.True(t, ok, "could not find datadog library init container", lang)
+		expectedVolumeMounts := []corev1.VolumeMount{
+			{
+				Name:      "datadog-auto-instrumentation",
+				MountPath: "/datadog-lib",
+				SubPath:   fmt.Sprintf("opt/datadog/apm/library/%s", lang),
+			},
+			{
+				Name:      "datadog-auto-instrumentation",
+				MountPath: "/opt/datadog-packages/datadog-apm-inject",
+				SubPath:   "opt/datadog-packages/datadog-apm-inject",
+			},
+		}
+		validator.requireVolumeMounts(t, expectedVolumeMounts)
+		validator.requireCommand(t, "/bin/sh -c -- sh copy-lib.sh /datadog-lib && echo $(date +%s) >> /opt/datadog-packages/datadog-apm-inject/c-init-time.datadog-lib-"+lang+"-init")
+	}
+}
+
+func (v *podValidator) requireAnnotations(t *testing.T, expected map[string]string) {
+	for key, expectedValue := range expected {
+		actualValue, exists := v.raw.Annotations[key]
+		require.True(t, exists, "annotation does not exist", key)
+		require.Equal(t, expectedValue, actualValue, "annotation does not match expected value")
+	}
+}
+
+func (v *podValidator) requireVolumeNames(t *testing.T, expected []string) {
+	require.Equal(t, len(expected), len(v.raw.Spec.Volumes), "expected volume count doesn't match actual volume count")
+	for i, volume := range v.raw.Spec.Volumes {
+		require.Equal(t, expected[i], volume.Name, "volume names do not match")
+	}
+}
+
+func (v *podValidator) requireEnvs(t *testing.T, expected map[string]string, expectedContainers []string) {
+	for _, name := range expectedContainers {
+		validator, found := v.containers[name]
+		require.True(t, found, "invalid test setup, expected container %s does not exist in the pod", name)
+		validator.requireEnvs(t, expected)
+	}
+}
+
+func (v *podValidator) requireMissingEnvs(t *testing.T, missing []string, expectedContainers []string) {
+	for _, name := range expectedContainers {
+		validator, found := v.containers[name]
+		require.True(t, found, "invalid test setup, expected container %s does not exist in the pod", name)
+		validator.requireMissingEnvs(t, missing)
+	}
+}
+
+func (v *podValidator) requireInjectorVersion(t *testing.T, expected string) {
+	require.Equal(t, expected, v.injectorVersion, "the injector version does not match the expected")
+}
+
+func (v *podValidator) requireLibraryVersions(t *testing.T, expected map[string]string) {
+	require.Equal(t, expected, v.libraryVersions, "the injected library versions do not match the expected")
+}
+
+func (v *podValidator) requireInitContainerImages(t *testing.T, expected []string) {
+	require.ElementsMatch(t, expected, v.initContainerImages, "init container images do not match expectated")
+}
+
+type containerValidator struct {
+	raw         *corev1.Container
+	image       *image
+	envs        map[string]string
+	commandline string
+}
+
+func newContainerValidator(container *corev1.Container) *containerValidator {
+	return &containerValidator{
+		raw:         container,
+		envs:        newEnvMap(container.Env),
+		image:       newImage(container.Image),
+		commandline: parseCommandline(container),
+	}
+}
+
+func (v *containerValidator) requireInjection(t *testing.T) {
+	expectedEnvs := map[string]string{
+		"LD_PRELOAD":            "/opt/datadog-packages/datadog-apm-inject/stable/inject/launcher.preload.so",
+		"DD_INJECT_SENDER_TYPE": "k8s",
+	}
+	v.requireEnvs(t, expectedEnvs)
+
+	expectedVolumeMounts := []corev1.VolumeMount{
+		{
+			Name:      "datadog-auto-instrumentation",
+			MountPath: "/opt/datadog-packages/datadog-apm-inject",
+			SubPath:   "opt/datadog-packages/datadog-apm-inject",
+		},
+		{
+			Name:      "datadog-auto-instrumentation-etc",
+			MountPath: "/etc/ld.so.preload",
+			SubPath:   "ld.so.preload",
+			ReadOnly:  true,
+		},
+		{
+			Name:      "datadog-auto-instrumentation",
+			MountPath: "/opt/datadog/apm/library",
+			SubPath:   "opt/datadog/apm/library",
+		},
+	}
+	v.requireVolumeMounts(t, expectedVolumeMounts)
+}
+
+func (v *containerValidator) requireNoInjection(t *testing.T) {
+	unsetEnvs := []string{
+		"LD_PRELOAD",
+		"DD_INJECT_SENDER_TYPE",
+		"DD_INSTRUMENTATION_INSTALL_TYPE",
+	}
+	v.requireMissingEnvs(t, unsetEnvs)
+	v.requireNoVolumeMounts(t)
+}
+
+func (v *containerValidator) requireVolumeMounts(t *testing.T, expected []corev1.VolumeMount) {
+	require.Equal(t, expected, v.raw.VolumeMounts, "volume mounts do not match expected")
+}
+
+func (v *containerValidator) requireCommand(t *testing.T, expected string) {
+	require.Equal(t, expected, v.commandline, "command line does not match expected")
+}
+
+func (v *containerValidator) requireSecurityContext(t *testing.T, expected *corev1.SecurityContext) {
+	require.Equal(t, expected, v.raw.SecurityContext, "security context not match expected")
+}
+
+func (v *containerValidator) requireResourceRequirements(t *testing.T, expected *corev1.ResourceRequirements) {
+	if expected == nil {
+		expected = &corev1.ResourceRequirements{
+			Limits:   corev1.ResourceList{},
+			Requests: corev1.ResourceList{},
+		}
+	}
+
+	require.Zero(t, expected.Requests.Memory().Cmp(*v.raw.Resources.Requests.Memory()), "expected memory request: %s, actual: %s", expected.Requests.Memory().String(), v.raw.Resources.Requests.Memory().String())
+	require.Zero(t, expected.Limits.Memory().Cmp(*v.raw.Resources.Limits.Memory()), "expected memory limit: %s, actual: %s", expected.Limits.Memory().String(), v.raw.Resources.Limits.Memory().String())
+	require.Zero(t, expected.Requests.Cpu().Cmp(*v.raw.Resources.Requests.Cpu()), "expected cpu request: %s, actual: %s", expected.Requests.Cpu().String(), v.raw.Resources.Requests.Cpu().String())
+	require.Zero(t, expected.Limits.Cpu().Cmp(*v.raw.Resources.Limits.Cpu()), "expected cpu limit: %s, actual: %s", expected.Limits.Cpu().String(), v.raw.Resources.Limits.Cpu().String())
+}
+
+func (v *containerValidator) requireNoVolumeMounts(t *testing.T) {
+	require.Empty(t, v.raw.VolumeMounts, "container should not have additional volume mounts")
+}
+
+func (v *containerValidator) requireEnvs(t *testing.T, expected map[string]string) {
+	for key, expectedValue := range expected {
+		actualValue, found := v.envs[key]
+		require.True(t, found, "could not find expected env %s in environment", key)
+		require.Equal(t, expectedValue, actualValue, "environment values do not match for container")
+	}
+}
+
+func (v *containerValidator) requireMissingEnvs(t *testing.T, missing []string) {
+	for _, key := range missing {
+		actualValue, found := v.envs[key]
+		require.False(t, found, "found %s in environment set to %s when it was not expected to be set", key, actualValue)
+	}
+}
+
+func newEnvMap(in []corev1.EnvVar) map[string]string {
+	envVars := make(map[string]string, len(in))
+	for _, env := range in {
+		envVars[env.Name] = env.Value
+	}
+	return envVars
+}
+
+func parseInjectorVersion(pod *corev1.Pod) string {
+	for _, container := range pod.Spec.InitContainers {
+		if container.Name == "datadog-init-apm-inject" {
+			parts := strings.Split(container.Image, ":")
+			if len(parts) != 2 {
+				continue
+			}
+			return parts[1]
+		}
+	}
+
+	return ""
+}
+
+func parseCommandline(container *corev1.Container) string {
+	command := strings.Join(container.Command, " ")
+	args := strings.Join(container.Args, " ")
+	return strings.Join([]string{command, args}, " ")
+}
+
+func difference(a, b []string) []string {
+	setB := make(map[string]bool, len(b))
+	for _, elem := range b {
+		setB[elem] = true
+	}
+
+	diff := []string{}
+	for _, elem := range a {
+		if _, found := setB[elem]; !found {
+			diff = append(diff, elem)
+		}
+	}
+
+	return diff
+}
+
+func parseAllContainerNames(pod *corev1.Pod) []string {
+	names := make([]string, len(pod.Spec.Containers))
+	for i, container := range pod.Spec.Containers {
+		names[i] = container.Name
+	}
+	return names
+}
+
+func parseInitContainerImages(pod *corev1.Pod) []string {
+	images := make([]string, len(pod.Spec.InitContainers))
+	for i, container := range pod.Spec.InitContainers {
+		images[i] = container.Image
+	}
+	return images
+}
+
+func parseLibraryVersions(pod *corev1.Pod) map[string]string {
+	injectedVersions := map[string]string{}
+
+	for _, container := range pod.Spec.InitContainers {
+		// gcr.io/datadoghq/dd-lib-java-init:v1
+		parts := strings.Split(container.Image, ":")
+		if len(parts) != 2 {
+			continue
+		}
+		fullImage := parts[0]
+		version := parts[1]
+
+		// gcr.io/datadoghq/dd-lib-java-init
+		parts = strings.Split(fullImage, "/")
+		if len(parts) < 1 {
+			continue
+		}
+		image := parts[len(parts)-1]
+
+		// dd-lib-java-init
+		parts = strings.Split(image, "-")
+		if len(parts) != 4 {
+			continue
+		}
+
+		// java
+		lib := parts[2]
+		injectedVersions[lib] = version
+	}
+
+	return injectedVersions
+}
+
+func newTestNamespace(name string, labels map[string]string) workloadmeta.KubernetesMetadata {
+	return workloadmeta.KubernetesMetadata{
+		EntityID: workloadmeta.EntityID{
+			Kind: workloadmeta.KindKubernetesMetadata,
+			ID:   string(util.GenerateKubeMetadataEntityID("", "namespaces", "", name)),
+		},
+		EntityMeta: workloadmeta.EntityMeta{
+			Name:   name,
+			Labels: labels,
+		},
+	}
+}
+
+func languageSetOf(languages ...string) languagemodels.LanguageSet {
+	set := languagemodels.LanguageSet{}
+	for _, l := range languages {
+		_ = set.Add(languagemodels.LanguageName(l))
+	}
+	return set
+}
+
+// MockImageResolver is a simple implementation that returns the original image unchanged.
+type MockImageResolver struct{}
+
+// newNoOpImageResolver creates a new noOpImageResolver.
+func NewFakeImageResolver() *MockImageResolver {
+	return &MockImageResolver{}
+}
+
+// Resolve returns the original image reference.
+func (r *MockImageResolver) Resolve(_ string, _ string, _ string) (*autoinstrumentation.ResolvedImage, bool) {
+	return nil, false
+}
